Create south_punjab_sentinel2
Modification to accompany the paper: 
Garcia-Molsosa, A. Orengo H.A. Petrie C.A. 'Reconstructing long-term settlement histories on complex alluvial floodplains: integrating historical map analysis and remote-sensing in the archaeological analysis of the landscape of the Indus River Basin'
Some changes have been applied in order to use Sentinel 2 instead of Landsat 5. The AOI has been adjusted to the paper study area.